+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1715] Fixed an issue where `Shape Changer` or `Mesh Cutter` could incorrectly delete all vertex color data in a mesh
 - [#1719] `Scale Adjuster` did not update positions of child objects when editing directly using the inspector fields
 - [#1726] Parameter drivers did not work properly when parameter types were adjusted after merging animators
+- [#1728] Menu Items that did not have a reactive component attached to their game object, but did have a child object 
+with a reactive component, would not function properly.
 
 ### Changed
 - [#1705] Reactive Component initial states are now applied on non-VRChat platforms